requirements.txt
django==1.10.5
django-grappelli==2.9.1
django-ckeditor==5.2.1
pillow==4.0.0
